Добавление проверки в Кафки (весь бренд в топике v1.brand)

+++ src/test/java/com/uplatform/wallet_tests/api/kafka/dto/core/gambling/v1/brand/payload/BrandDeletePayload.java
@@ -14,5 +14,5 @@ public class BrandDeletePayload {
     private String uuid;
 
     @JsonProperty("deleted_at")
-    private Long deleted_at;
+    private Integer deleted_at;
 }
