+++ src/test/java/com/uplatform/wallet_tests/api/kafka/dto/core/gambling/v1/brand/payload/BrandUpdatePayload.java
@@ -29,5 +29,5 @@ public class BrandUpdatePayload {
     private Boolean status_enabled;
 
     @JsonProperty("updated_at")
-    private Long updated_at;
+    private Integer updated_at;
 }
